requirements.txt
# Runtime dependencies (validators and skills)
pyyaml>=6.0

# Test dependencies
pytest>=8.0
